requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.0.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
docopt==0.6.2
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
Js2Py==0.74
packaging==23.2
pipwin==0.5.2
psycopg2==2.9.9
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.1.0
